+++ frontendErgoBalance/src/components/Header/header.scss
@@ -1,0 +1,36 @@
+.header__wrapper {
+    width: 100vw;
+    height: 5rem;
+    position: sticky;
+    top: 0;
+    margin-bottom: 0.5rem;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+}
+
+.header__logo {
+    z-index: 99;
+    height: 3rem;
+    background-color: white;
+    padding: 2px;
+    margin-top: 0.7rem;
+    margin-left: 1.5rem;
+    border-radius: 100%;
+    cursor: pointer;
+
+    &:hover {
+        scale: 1.1;
+    }
+
+    &:active {
+        scale: 0.9;
+    }
+}
+
+.header__btn {
+    margin-top: 0.7rem;
+    margin-right: 1.5rem;
+    background-color: rgb(47, 47, 47) !important;
+    border: none !important;
+}
