+++ frontendErgoBalance/src/views/aboutUsView/aboutUsView.scss
@@ -17,23 +17,6 @@
   background-color: var(--color-white) !important;
 }
 
-.aboutUsView__logo {
-  z-index: 99;
-  height: 3rem;
-  position: absolute;
-  top: 0.5rem;
-  left: 0.5rem;
-  cursor: pointer;
-
-  &:hover {
-    scale: 1.1;
-  }
-
-  &:active {
-    scale: 0.9;
-  }
-}
-
 .aboutUsView__header {
   text-transform: uppercase;
   font-size: 1.5rem;
