+++ frontendErgoBalance/src/views/intervalView/intervalView.scss
@@ -19,23 +19,6 @@
     position: relative;
 }
 
-.intervalView__logo {
-    z-index: 10;
-    height: 3rem;
-    position: absolute;
-    top: 0.5rem;
-    left: 0.5rem;
-    cursor: pointer;
-
-    &:hover {
-        scale: 1.1;
-    }
-
-    &:active {
-        scale: 0.9;
-    }
-}
-
 .active-view {
     display: flex;
     flex-direction: column;
