+++ frontendErgoBalance/src/views/setupExercisesView/setupExercisesView.scss
@@ -2,27 +2,8 @@
 
 .setupExercisesView__wrapper {
     @include wrapper;
-    background: rgb(162,208,193) !important;
-    background: linear-gradient(144deg, rgba(162,208,193,1) 0%, rgba(11,88,80,1) 60%) !important; 
-}
-
-.setupExercisesView__logo {
-    z-index: 10;
-    height: 3rem;
-    position: absolute;
-    top: 0.5rem;
-    left: 0.5rem;
-    cursor: pointer;
-    background-color: var(--color-white);
-    border-radius: 50%;
-    padding: .1rem;
-    &:hover {
-        scale: 1.1;
-    }
-    
-    &:active {
-        scale: 0.9;
-    }
+    background: rgb(162, 208, 193) !important;
+    background: linear-gradient(144deg, rgba(162, 208, 193, 1) 0%, rgba(11, 88, 80, 1) 60%) !important;
 }
 
 .setupExercisesView__title {
@@ -31,7 +12,7 @@
 
 .setupExercisesView__container {
     @include centering;
-    background-color: rgba(255,255,255,.3);
+    background-color: rgba(255, 255, 255, .3);
     color: var(--color-white);
     border-radius: #{$borders-radius-md};
     margin: 1rem;
@@ -58,11 +39,14 @@
 .setupExercisesView__li--inner {
     cursor: pointer;
 }
+
 .setupExercisesView__button {
     @include button-base;
     margin: 2rem .5rem;
 }
-.setupExercisesView__button.setupExercisesView__button--remove, .setupExercisesView__button.setupExercisesView__button--add {
+
+.setupExercisesView__button.setupExercisesView__button--remove,
+.setupExercisesView__button.setupExercisesView__button--add {
     border-color: var(--color-black) !important;
     min-width: 30px !important;
     max-width: 30px !important;
@@ -81,7 +65,7 @@
     width: 100vw;
     height: 100vh;
     background: var(--color-black);
-    background: linear-gradient(144deg, rgba(162,208,193,1) 0%, rgba(11,88,80,1) 60%) !important; 
+    background: linear-gradient(144deg, rgba(162, 208, 193, 1) 0%, rgba(11, 88, 80, 1) 60%) !important;
     padding: #{$spacing-xl};
 }
 
@@ -98,6 +82,7 @@
     font-size: #{$sizing-xs};
     padding: #{$spacing-md};
 }
+
 .setupExercisesView__modal-text {
     font-size: 1rem;
     padding: #{$spacing-md};
@@ -113,9 +98,10 @@
     top: 0;
     right: 1rem;
 }
+
 .setupExercisesView__loading {
     @include centering;
-    background-color: rgba(255,255,255,.3);
+    background-color: rgba(255, 255, 255, .3);
     min-width: 90%;
     color: var(--color-white);
     border-radius: #{$borders-radius-md};
@@ -127,13 +113,14 @@
     flex-direction: column;
     justify-content: center;
     align-items: center;
-    gap:1rem;
+    gap: 1rem;
 }
 
 .loading__image {
     width: 280px;
     height: 280px;
 }
+
 .p-progressspinner-circle {
     stroke: var(--color-black) !important;
 }
